requirements.txt
asgiref==3.2.7
Babel==2.8.0
Django==3.0.6
django-isbn-field==0.5.2
django-phonenumber-field==4.0.0
djangorestframework==3.11.0
future==0.18.2
phonenumbers==8.12.5
python-stdnum==1.13
pytz==2020.1
six==1.15.0
sqlparse==0.3.1
